+++ src/commonMain/kotlin/com/codeancy/metroui/domain/models/LocationInfoedUi.kt
@@ -2,5 +2,5 @@ package com.codeancy.metroui.domain.models
 
 data class LocationInfoedUi(
     val location: LocationUi,
-    val entity: Int
+    val entity: Long
 )
