+++ src/commonMain/kotlin/com/codeancy/metroui/route/utils/RouteScreenRouteUi.kt
@@ -1,7 +1,7 @@
 package com.codeancy.metroui.route.utils
 
 data class RouteScreenRouteUi(
-    val sourceId: Int,
-    val destId: Int,
+    val sourceId: Long,
+    val destId: Long,
     val isRecent: Boolean = false
 )
